requirements.txt
tensorflow==2.6.2
scipy==1.5.4
numpy==1.19.5
PyYAML==6.0
